+++ packages/grafana-ui/src/components/DataLinks/codemirrorUtils.ts
@@ -6,10 +6,10 @@ import { DataLinkBuiltInVars, type GrafanaTheme2, VariableOrigin, type VariableS
 import { type CodeMirrorCompletionContext, type CodeMirrorCompletionResult } from '../CodeMirror/types';
 import { createVariableCompletionSource } from '../CodeMirror/variableCompletion';
 
-export type DataLinkInterpolationMode = 'url' | 'text';
-
 // Matches a complete `${...}` variable reference anywhere in the document.
 const VARIABLE_PATTERN = /\$\{[^}]+\}/g;
+
+export type DataLinkInterpolationMode = 'url' | 'text';
 
 const VARIABLE_CLASS = 'cm-variable';
