+++ public/app/features/notebook/scene/layout-notebook/NotebookCellRenderer.tsx
@@ -1,4 +1,5 @@
 import { css } from '@emotion/css';
+import { Suspense } from 'react';
 
 import { type GrafanaTheme2 } from '@grafana/data';
 import { type VizPanel } from '@grafana/scenes';
@@ -78,12 +79,14 @@ function NarrativeCell({
   const Renderer = registered.render;
   return (
     <div className={styles.content}>
-      <Renderer
-        content={content}
-        isEditing={isEditing}
-        autoFocus={autoFocus}
-        onChange={(updated) => cell.onContentChange(updated)}
-      />
+      <Suspense fallback={content.kind === 'Code' ? <pre>{content.spec.code}</pre> : null}>
+        <Renderer
+          content={content}
+          isEditing={isEditing}
+          autoFocus={autoFocus}
+          onChange={(updated) => cell.onContentChange(updated)}
+        />
+      </Suspense>
     </div>
   );
 }
